Console: Make uart a module of console
Uart is going to be a module of console since Console should be the only
interface to the UART other than the initialization. Console will be
stateless and be accessed through the print! and println! macros which
will be specified in a later commit.

This probably doesn't compile on its own.