Console: Implement Console::read() wrapper around uart
Console::read() wraps the uart rx register. In its current form, the
input is buffered until the user presses their enter key, indicating the
end of the line. Additionally, the buffer has a maximum of 256
characters. Eventually, this can be changed to grow dynamically. The
input supplied by the user is currentlys shown to the screen and then
echoed again from lib.rs where the Console::read() call was made to
indicate the character array is successfully returned.

We now have basic i/o capabilities in our OS :)