start.S: Add trap handler setup code
The code added here sets up the state of the OS for handling traps. The
trap_handler label will save the current register context (i.e. saving
the mepc register) and put mcause as the argument for the rust function
that will be called to match the actual cause of the trap.