start.S: Add trap_handler label
Add label to load in mtvec for handling traps, interrupts, and
exceptions during runtime. When any of the above are caught, the current
execution will be stopped and control will be given to that assembly
label where the current context must be saved and then control given to
rust where a stronger handler will parse mcause (maybe needs more
information) and call the proper handlers.