Trap: Add rust trap handler function and test
The rust trap handler will need to check the value of the mcause
register which is passed in from the assembly trap point and then call
the proper handler for the cause of the trap. Currently, this spins in a
loop, and I'm not entirely sure why. We can receive mcause = 0x02 which
is an illegal instruction fault, but nothing else for some reason.