Interrupt: Add explicit initialization of interrupts
Before, interrupts were enabled in the bootstrapping of the first
processor, at the same time machine mode was enabled. Not having the
guarantee of being in machine mode makes this action a potential issue
during the booting procedure. This patch moves the enabling of
interrupts as an explicit call after rust code has started executing.
This simply enables all interrupts in the mstatus register using the MIE
and MPIE bits. This is also where mtvec register is set to the trap
handling function where {traps, interrupts, exceptions} will be sent
initially and dispersed to the proper handling routines.