Timer: Add timer initialization
Initialize the timer to occur once per second - which for the HiFive
board is once every ~17.5 million clock ticks. The timer will trigger
after one second and will then enter and endless loop since the
comparison timer is never increased to be higher than the actual value
of the clock again. This simply gets the timer initialized and makes
sure we receive the interrupt.

Since the timer is already defined in src/console/uart.rs, that will
just be a public definition until we put all the constants like that in
their own file at the top level or something.