Context: Create struct to store contexts
This struct will be the main abstraction for handling and dealing with
context switching. Since this is a pretty involved task, this interface
will probably be drastically changing as more contexts are added. This
commit will serve as a starting point for the design of the struct and
its implementation.