ADVOS: Move global, board specific constants to a file
At the top level, there is a module called global_constants which can
now be used in all submodules of the crate in the form of
`use crate::global_constants::{CONST_NAME}`. This provides a uniform
way for submodules to access these constants rather than just keeping
them in the first file they're needed.