Timer: Now works as expected
Per Dr. Marz' suggestion, the mtime_cmp_{hi,lo} values should be
incremented individually of eachother instead of shoving them both into
a u64. Once they are incremented with the constant offset, they are
rewritten to the CLINT.

Fixed a bug where mtime_cmp_lo was being written to with the value for
time_cmp_lo and then written again with the value of time_cmp_hi.