Format: Add rustfmt.toml and format all Rust code
rustfmt gives us an easy way to format our code the same and for the
compiler to not complain to us about formatting. I have a pre-commit
hook that runs cargo fmt (which formats the entire crate) instead of
running rustfmt on each file.

Now that everything is formatted properly, we should only have to format
new code.