Testing: Clean up testing procedures in lib.rs
Removed most of the print statements from the tests and replaced them
with assert! and assert_eq!. The tests no longer print much information
but will panic on failure. More printing can be added if necessary.

When `make test` is run, the kernel should not panic, and should enter
an infinite loop after a message to quit qemu has been printed.