Scheduler: Add GLOBAL_CTX from assembly
To access GLOBAL_CTX, which is defined in assembly to be the start of
the bss section, we need a static global variable that will be the
pointer to that region of memory. Since it's defined outside of Rust, it
must be in an extern "C" block which is inherently unsafe.