Scheduler: Add something that kind of works maybe
This should work, I just don't know how to make the scheduler object
available to the interrupt handler. Our options seem to be making a
scheduler object globally available, which would be availablet to the
interrupt handler, or having a handle to a singleton method Scheduler
which would have to pass references around to, but since you can't pass
the reference to the interrupt handler you have to make the function
static (without any kind of reference to self, which implies that we
won't have access to the process list) I'm stuck.