Scheduler: Add a _lot_ more implementation
The scheduler is coming along. Currently iterate between two processes
(the scheduler and one other process at the moment). This patch adds the
main scheduling algorithm that gets bootstrapped and can create processes.
The next things that will need to be added is correct deallocation and cleanup
of the process list.