Trap: Add acquiring lock logic and print counter
The scheduler now has a lock to make sure that the process list doesn't
get edited by multiple processes at the same time or while one is
preempted by another. The timer will try to acquire the lock, but if it
cannot it will simply wait until the next interrupt to try again.