Scheduler: Fixup the process deletion flow
Instead of deallocating processes when they exit, we'll just leave them
in the process list to be reused. This is needed because we can't remove
processes in the middle of the HeapVec right now in constant time
without leaving "empty" spaces, which will almost certainly cause issues
later. Now the create_proc method will check from the beginning of the
list to find a process that has ProcessState::Exited and will simply
overwrite that one.