Trap: Fix off by one error in register macro
The macros were placing the registers from the context of a PCB
structure one off the correct placement, probably because the PCB struct
held all 32 registers and the macros were only working with 31 because I
was trying to be clever and not do anything with x0. Oh well, that's
fixed now.