Syscalls: Add the SyscallTable and ecall wrapper
The SyscallTable will be where we all system calls will be identified
with a number. So far, I have put the ones that I think we _definitely_
need to implement, and there's definitely room for more if we want them.

The ecall wrapper will be called by every system call with the
appropriate parameters and will trigger the trap to perform the actual
operations which will eventually call _{syscall} so for example when
exit(0) is called, the only thing the exit() function will do is call
ecall(SycallTable::EXIT, 0) and when the ecall is caught, based on the
value of SyscallTable parameter we will call _exit(0) to actually exit
the process.