Scheduler: Simplify Scheduler::create_proc
The process creation strategy is something that will have to be
developed with care since we don't want to have to deallocate processes
all the time and we don't want to have our HeapVec become fragmented.
Right now I just push on to the heapvec and they are never removed.